build: refactor cert file creation for CI, so it can be executed in all modules
